refactor(weather): Move weather services to data layer
- Relocated API client implementations to lib/data/services/weather/services/
- Updated import paths to reflect separation of concerns
- Supports better modularization aligned with Clean Architecture
